Removed files due to successful refractoring, working on the refactoring of StartLocationSetting, implemented pagination
LocationSelection2 has been broken down into to files in BaseLocation directory (LocationSelection, RadiusSet, SearchInitialLocation)

DestinationList2 has been refractored to files in DestinationList folder
