+++ Frontend/TrekReact/screens/GameScreens/GameNavigator.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import { createStackNavigator } from '@react-navigation/stack';
-import App from './LocationSelection';
-import NearbyDestinationsScreen from './DestinationList/DestinationList2'; // Import your second code component
+import NearbyDestinationsScreen from './DestinationList/DestinationList'; // Import your second code component
 import SelectStartLocationScreen from './StartLocationSetting';
 import StartGameScreen from './StartGameUI';
 import GameMapScreen from './MainGameUI';
-import LocationSelectionScreen from '../GameScreens/BaseLocation/LocationSelection2';
+import LocationSelectionScreen from '../GameScreens/BaseLocation/LocationSelection';
 import RadiusSetScreen from '../GameScreens/BaseLocation/RadiusSet';
 import SearchLocationScreen from '../GameScreens/BaseLocation/SearchInitialLocation';
 const Stack = createStackNavigator();
@@ -32,14 +31,6 @@ function GameNavigation() {
       <Stack.Screen
         name="SearchLocation"
         component={SearchLocationScreen}
-        options={{ 
-          headerShown: false, 
-          gestureEnabled: false 
-        }} 
-      />
-      <Stack.Screen
-        name="Location Selection"
-        component={App}
         options={{ 
           headerShown: false, 
           gestureEnabled: false
