+++ Frontend/TrekReact/screens/GameScreens/StartGameUI.js
@@ -1,3 +1,5 @@
+
+
 import React from 'react';
 import { View, StyleSheet, Text, TouchableOpacity, Alert } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
